+++ CHANGELOG.md
@@ -8,6 +8,8 @@ ENHANCEMENTS:
 
 BUG FIXES:
 
+* Changes in `create_before_destroy` for resources which require replacement are now properly handled when refresh is disabled. ([#2248](https://github.com/opentofu/opentofu/pull/2248))
+
 ## Previous Releases
 
 For information on prior major and minor releases, see their changelogs: