+++ CHANGELOG.md
@@ -9,6 +9,8 @@ ENHANCEMENTS:
 BUG FIXES:
 * `tofu init` command does not attempt to read encryption keys when `-backend=false` flag is set. (https://github.com/opentofu/opentofu/pull/2293)
 
+* Changes in `create_before_destroy` for resources which require replacement are now properly handled when refresh is disabled. ([#2248](https://github.com/opentofu/opentofu/pull/2248))
+
 ## Previous Releases
 
 For information on prior major and minor releases, see their changelogs: